refactor: 위치 선택 기능 공통 컴포넌트로 분리 (#232)
