+++ apps/web/components/location/index.ts
@@ -1,7 +1,8 @@
 export { default as AddressSearchButton } from './AddressSearchButton';
+export { default as CurrentLocationButton } from './CurrentLocationButton';
 export { default as LocationDisplay } from './LocationDisplay';
 export { default as LocationHeader } from './LocationHeader';
 export { default as LocationMap } from './LocationMap';
 export { default as LocationMapContainer } from './LocationMapContainer';
+export { default as LocationSelector } from './LocationSelector';
 export { default as SaveLocationButton } from './SaveLocationButton';
-export { default as CurrentLocationButton } from './CurrentLocationButton';
